refactor: complete migration to CSS export and remove Less build pipeline
- replace root all.less with cumulative all.css entry point
- update website style imports to consume all.css
- consolidate CSS style worker into single worker/css.mjs and remove Less workers
- remove remaining Less configuration, targets, and documentation references in esl package
