+++ packages/esl-website/src/lib.less
@@ -2,7 +2,7 @@
 @import './common/typography.less';
 @import './common/colors.less';
 
-@import '@exadel/esl/modules/all.less';
+@import (inline) '@exadel/esl/modules/all.css';
 
 @import './common/accent-lines.less';
 @import './common/accordion.less';
