+++ packages/esl-website/src/site.less
@@ -1,4 +1,4 @@
-@import '@exadel/esl/modules/all.less';
+@import (inline) '@exadel/esl/modules/all.css';
 
 @import './common/reboot.less';
 @import './common/typography.less';
