+++ packages/esl/package.json
@@ -33,8 +33,8 @@
     "scope": "@exadel"
   },
   "files": [
-    "modules/**/*.{js,ts,less,css,md}",
-    "polyfills/**/*.{js,ts,less,css,md}",
+    "modules/**/*.{js,ts,css,md}",
+    "polyfills/**/*.{js,ts,css,md}",
     "skills/**/*.md",
     "tsconfig.json",
     "CHANGELOG.md",
