+++ packages/esl/project.json
@@ -65,28 +65,22 @@
     "build:css": {
       "executor": "nx:run-commands",
       "options": {
-        "commands": [
-          "job-ripper \"src/**/*.less\" -w worker/less.mjs -c 2",
-          "job-ripper \"src/**/*.css\" -w worker/css.mjs -c 2"
-        ],
-        "parallel": true,
+        "command": "job-ripper \"src/**/*.css\" -w worker/css.mjs -c 2",
         "cwd": "packages/esl"
       },
       "cache": true,
       "inputs": [
         "{projectRoot}/src/**/*.css",
-        "{projectRoot}/src/**/*.less",
         "{projectRoot}/worker/**/*.mjs"
       ],
       "outputs": [
-        "{projectRoot}/modules/**/*.css",
-        "{projectRoot}/modules/**/*.less"
+        "{projectRoot}/modules/**/*.css"
       ]
     },
     "watch:css": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "chokidar \"src/**/*.{less,css}\" -c \"nx run esl:build:css\"",
+        "command": "chokidar \"src/**/*.css\" -c \"nx run esl:build:css\"",
         "cwd": "packages/esl"
       }
     },
@@ -124,12 +118,11 @@
     "lint:stylelint": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "stylelint \"src/**/*.{less,css}\"",
+        "command": "stylelint \"src/**/*.css\"",
         "cwd": "packages/esl"
       },
       "cache": true,
       "inputs": [
-        "{projectRoot}/src/**/*.less",
         "{projectRoot}/src/**/*.css",
         "{projectRoot}/.stylelint.yml"
       ]
