--- packages/esl/src/all.less
@@ -1,61 +1,0 @@
-// All in one bundle
-@import (inline) './esl-utils/dom/scroll.css';
-
-@import (inline) './esl-image-utils/core/esl-image-utils.container.css';
-@import (inline) './esl-image-utils/core/esl-image-utils.fade.css';
-@import (inline) './esl-image-utils/core/esl-image-utils.ratios.css';
-
-@import (inline) './esl-media/core/esl-media.css';
-@import (inline) './esl-media/core/esl-media-container.css';
-
-@import (inline) './esl-scrollbar/core/esl-scrollbar.css';
-@import (inline) './esl-scrollbar/core/esl-scrollbar-arrows.css';
-@import (inline) './esl-scrollbar/core/esl-scrollable-content.css';
-
-// CSS entry points retain @import rules when inlined by Less. During the CSS
-// migration, include their leaves directly so the all-in-one output is bundled.
-@import (inline) './esl-tab/core/esl-tab.css';
-@import (inline) './esl-tab/core/esl-tabs.css';
-@import (inline) './esl-tab/core/esl-tabs.scrollable.css';
-@import (inline) './esl-tab/core/esl-tabs.arrows.css';
-@import (inline) './esl-toggleable/core/esl-toggleable.css';
-@import (inline) './esl-toggleable/core/esl-toggleable-dispatcher.css';
-@import (inline) './esl-trigger/core/esl-trigger.css';
-@import (inline) './esl-panel/core/esl-panel.css';
-@import (inline) './esl-panel-group/core/esl-panel-group.css';
-
-@import (inline) './esl-popup/core/esl-popup.css';
-@import (inline) './esl-tooltip/core/esl-tooltip.css';
-@import (inline) './esl-footnotes/core/esl-footnotes.css';
-@import (inline) './esl-footnotes/core/esl-note.css';
-@import (inline) './esl-share/core/esl-share.css';
-@import (inline) './esl-share/core/esl-share-list.css';
-@import (inline) './esl-share/core/esl-share-button.css';
-
-@import (inline) './esl-anchornav/core/esl-anchor.css';
-@import (inline) './esl-anchornav/core/esl-anchornav.css';
-@import (inline) './esl-anchornav/core/esl-anchornav-sticked.css';
-@import (inline) './esl-a11y-group/core/esl-a11y-group.css';
-@import (inline) './esl-alert/core/esl-alert.css';
-@import (inline) './esl-animate/core/esl-animate.css';
-@import (inline) './esl-animate/core/esl-animate-animations.css';
-@import (inline) './esl-drag-to-scroll/core/esl-drag-to-scroll.css';
-@import (inline) './esl-line-clamp/core/esl-line-clamp.css';
-@import (inline) './esl-sticky-box/core/esl-sticky-box.css';
-
-/* './esl-carousel/all.css' */
-@import (inline) './esl-carousel/core/esl-carousel.css';
-@import (inline) './esl-carousel/renderers/esl-carousel.default.renderer.css';
-@import (inline) './esl-carousel/renderers/esl-carousel.grid.renderer.css';
-@import (inline) './esl-carousel/renderers/esl-carousel.css.renderer.css';
-@import (inline) './esl-carousel/renderers/esl-carousel.css.fade.renderer.css';
-@import (inline) './esl-carousel/renderers/esl-carousel.css.slide.renderer.css';
-@import (inline) './esl-carousel/plugin/nav/esl-carousel.nav.arrows.css';
-@import (inline) './esl-carousel/plugin/dots/esl-carousel.nav.dots.css';
-
-/* './esl-forms/all.css' */
-@import (inline) './esl-forms/esl-select/core/esl-select.css';
-@import (inline) './esl-forms/esl-select/core/esl-select-renderer.css';
-@import (inline) './esl-forms/esl-select/core/esl-select-dropdown.css';
-@import (inline) './esl-forms/esl-select-list/core/esl-select-item.css';
-@import (inline) './esl-forms/esl-select-list/core/esl-select-list.css';
