+++ packages/esl/worker/css.mjs
@@ -1,18 +1,25 @@
+import fs from 'node:fs/promises';
 import path from 'node:path';
 import postcss from 'postcss';
 import postcssImport from 'postcss-import';
-import {createStyleWorker} from './styles.mjs';
 
+const SRC_DIR = path.resolve('src');
+const OUT_DIR = path.resolve('modules');
 const processor = postcss([postcssImport()]);
 
-export default createStyleWorker({
-  ext: 'css',
-  copy: true,
-  compile: async (source, absFilePath) => {
-    const result = await processor.process(source, {from: absFilePath, map: false});
-    return result.css;
-  },
-  filter: (relFilePath) => {
-    return ['all.css', 'core.css'].includes(path.basename(relFilePath));
-  }
-});
+export default async function (filePath) {
+  const absFilePath = path.resolve(filePath);
+  const relFilePath = path.relative(SRC_DIR, absFilePath);
+  const destFilePath = path.join(OUT_DIR, relFilePath);
+
+  await fs.mkdir(path.dirname(destFilePath), {recursive: true});
+  await fs.copyFile(filePath, destFilePath);
+
+  if (!['all.css', 'core.css'].includes(path.basename(relFilePath))) return;
+
+  const source = await fs.readFile(filePath, 'utf-8');
+  const result = await processor.process(source.toString(), {from: absFilePath, map: false});
+
+  await fs.writeFile(destFilePath, result.css);
+}
+
